Add instance_max_run_duration_seconds config
This pr adds a new `instance_max_run_duration_seconds` config field
that gets piped down into the builder instance's `maxRunDuration`
schedule. This will allow for an instance to automatically be stopped
or deleted, even if the packer build process is killed.

It also adds an `instance_termination_action` which must be set if
`instance_max_run_duration_seconds` config is set. This will tell GCP
to either `STOP` or `DELETE` the instance once the max .