Add maxRuntimeDurationSeconds config
This pr adds a new `maxRuntimeDurationSeconds` config that gets piped
down into the builder instance's `maxRuntimeDuration` schedule. This will allow
for an instance to automatically be deleted even if the packer build is
interrupted.

It also adds a `instanceTerminationAction` which must be set if
`maxRuntimeDurationSeconds` config is set. This will tell GCP to either
`STOP` or `DELETE` the instance once it is terminated.